moved MockNetworkManager to a separate file
