+++ Countries/NetworkManager.swift
@@ -30,16 +30,3 @@ class NetworkManager: NetworkManagerProtocol {
     }
 }
 
-class MockNetworkManager: NetworkManagerProtocol {
-    
-    func getCountry() -> CountriesApiQuery.Data.Country {
-        let continent = CountriesApiQuery.Data.Country.Continent(name: "testcontinent")
-        let language = CountriesApiQuery.Data.Country.Language(name: "testlanguage")
-        return CountriesApiQuery.Data.Country(code: "UA", name: "testname", capital: "testcapital", currency: "testcurrency", phone: "testphone", continent: continent, languages: [language])
-    }
-    
-    func getCountries(completion: @escaping (Result<[CountriesApiQuery.Data.Country]?, Error>) -> Void) {
-        completion(Result.success([getCountry(), getCountry()]))
-    }
-}
-
